Refactor CDB firmware APIs to use CdbFwHandler
