+++ sonic_platform_base/sonic_xcvr/api/public/cmis.py
@@ -7,18 +7,12 @@
 
 from enum import Enum
 from ...fields import consts
-from ...fields import cdb_consts
 from ..xcvr_api import XcvrApi
-
+from .cdb import CmisCdbFw
 import logging
 from ...codes.public.cmis import CmisCodes
-from ...codes.public.cdb import CdbCodes
 from ...codes.public.sff8024 import Sff8024
 from ...fields import consts
-from ...mem_maps.public.cdb import CdbMemMap
-from ...cdb.cdb_fw import CdbFwHandler as CdbFw
-from ..xcvr_api import XcvrApi
-from .cmisCDB import CmisCdbApi
 from .cmisVDM import CmisVdmApi
 import time
 import copy
@@ -110,7 +104,7 @@ CMIS_XCVR_INFO_DEFAULT_DICT = {
         "vdm_supported": "N/A"
         }
 
-class CmisApi(XcvrApi):
+class CmisApi(CmisCdbFw, XcvrApi):
     NUM_CHANNELS = 8
     LowPwrRequestSW = 4
     LowPwrAllowRequestHW = 6
@@ -142,28 +136,7 @@ class CmisApi(XcvrApi):
     def __init__(self, xcvr_eeprom, init_cdb_fw_handler=False):
         super(CmisApi, self).__init__(xcvr_eeprom)
         self.vdm = CmisVdmApi(xcvr_eeprom) if not self.is_flat_memory() else None
-        self.cdb = CmisCdbApi(xcvr_eeprom) if self.is_cdb_supported() else None
-        self._init_cdb_fw_handler = init_cdb_fw_handler
-        self._cdb_fw_hdlr = None
-
-    @property
-    def cdb_fw_hdlr(self):
-        if not self._init_cdb_fw_handler:
-            return None
-        
-        if self._cdb_fw_hdlr is None:
-            self._cdb_fw_hdlr = self._create_cdb_fw_handler()
-        return self._cdb_fw_hdlr
-
-    def _create_cdb_fw_handler(self):
-        if not self.is_cdb_supported():
-            self._init_cdb_fw_handler = False
-            return None
-        cdb_mem_map = CdbMemMap(CdbCodes)
-        return CdbFw(self.xcvr_eeprom.reader, self.xcvr_eeprom.writer, cdb_mem_map)
-
-    def get_cdb_fw_handler(self):
-        return self.cdb_fw_hdlr
+        self._init_cdb(xcvr_eeprom, init_cdb_fw_handler)
 
     def _get_vdm_key_to_db_prefix_map(self):
         return CMIS_VDM_KEY_TO_DB_PREFIX_KEY_MAP
@@ -1693,22 +1666,6 @@ class CmisApi(XcvrApi):
         logger.error('Invalid loopback mode:%s, lane_mask:%#x', loopback_mode, lane_mask)
         return False
 
-    def is_cdb_supported(self):
-        '''
-        This function returns whether CDB is supported
-        '''
-        if self.is_flat_memory():
-            return False
-
-        cdb_inst = self.xcvr_eeprom.read(consts.CDB_SUPPORT)
-        if cdb_inst is None:
-            return False
-
-        if cdb_inst == 1 or cdb_inst == 2:
-            return True
-
-        return False
-
     @read_only_cached_api_return
     def is_transceiver_vdm_supported(self):
         '''
@@ -1837,443 +1794,6 @@ class CmisApi(XcvrApi):
                        'aux3_flags': aux3_flags,
                        'custom_mon_flags': custom_mon_flags}
         return module_flag
-
-    def get_module_fw_mgmt_feature(self, verbose = False):
-        """
-        This function obtains CDB features supported by the module from CDB command 0041h,
-        such as start header size, maximum block size, whether extended payload messaging
-        (page 0xA0 - 0xAF) or only local payload is supported. These features are important because
-        the following upgrade with depend on these parameters.
-        """
-        txt = ''
-        if self.cdb is None:
-            return {'status': False, 'info': "CDB Not supported", 'feature': None}
-
-        # get fw upgrade features (CMD 0041h)
-        starttime = time.time()
-        autopaging = self.xcvr_eeprom.read(consts.AUTO_PAGING_SUPPORT)
-        autopaging_flag = bool(autopaging)
-        writelength_raw = self.xcvr_eeprom.read(consts.CDB_SEQ_WRITE_LENGTH_EXT)
-        if writelength_raw is None:
-            return None
-        writelength = (writelength_raw + 1) * 8
-        txt += 'Auto page support: %s\n' %autopaging_flag
-        txt += 'Max write length: %d\n' %writelength
-        result = self.cdb.get_fw_management_features()
-        status = result['status']
-        _, rpl_chkcode, rpl = result['rpl']
-
-        if status == 1 and self.cdb.cdb_chkcode(rpl) == rpl_chkcode:
-            startLPLsize = rpl[2]
-            txt += 'Start payload size %d\n' % startLPLsize
-            maxblocksize = (rpl[4] + 1) * 8
-            txt += 'Max block size %d\n' % maxblocksize
-            lplEplSupport = {0x00 : 'No write to LPL/EPL supported',
-                            0x01 : 'Write to LPL supported',
-                            0x10 : 'Write to EPL supported',
-                            0x11 : 'Write to LPL/EPL supported'}
-            txt += '{}\n'.format(lplEplSupport[rpl[5]])
-            if rpl[5] == 1:
-                lplonly_flag = True
-            else:
-                lplonly_flag = False
-            txt += 'Abort CMD102h supported %s\n' %bool(rpl[1] & 0x01)
-            if verbose:
-                txt += 'Copy CMD108h supported %s\n' %bool((rpl[1] >> 1) & 0x01)
-                txt += 'Skipping erased blocks supported %s\n' %bool((rpl[1] >> 2) & 0x01)
-                txt += 'Full image readback supported %s\n' %bool((rpl[1] >> 7) & 0x01)
-                txt += 'Default erase byte {:#x}\n'.format(rpl[3])
-                txt += 'Read to LPL/EPL {:#x}\n'.format(rpl[6])
-
-        else:
-            txt += 'Status or reply payload check code error\n'
-            logger.error(txt)
-            logger.error('Fail to get fw mgmt feature, cdb status: {:#x}, cdb_chkcode: {:#x}, rpl_chkcode: {:#x}\n'.format(status, self.cdb.cdb_chkcode(rpl), rpl_chkcode))
-            return {'status': False, 'info': txt, 'feature': None}
-        elapsedtime = time.time()-starttime
-        logger.info('Get module FW upgrade features time: %.2f s\n' %elapsedtime)
-        logger.info(txt)
-        return {'status': True, 'info': txt, 'feature': (startLPLsize, maxblocksize, lplonly_flag, autopaging_flag, writelength)}
-
-    def get_module_fw_info(self):
-        """
-        This function returns firmware Image A and B version, running version, committed version
-        and whether both firmware images are valid.
-        Operational Status: 1 = running, 0 = not running
-        Administrative Status: 1=committed, 0=uncommitted
-        Validity Status: 1 = invalid, 0 = valid
-        """
-        txt = ''
-
-        if self.cdb is None:
-            return {'status': False, 'info': "CDB Not supported", 'result': None}
-
-        # Create CDB FW handler with exception handling
-        if self._cdb_fw_hdlr is None:
-            try:
-                self._cdb_fw_hdlr = self._create_cdb_fw_handler()
-            except Exception as e:
-                logger.error("CDB FW handler init assertion failed: %s", e)
-                return {'status': False, 'info': "CDB FW handler init failed", 'result': None}
-            if self._cdb_fw_hdlr is None:
-                return {'status': False, 'info': "CDB FW handler init failed", 'result': None}
-
-        fw_info = self._cdb_fw_hdlr.get_firmware_info()
-
-        # password issue
-        if fw_info is False or fw_info is None:
-            status = self.cdb.cdb1_chkstatus()
-            if status == 0x46:
-                logger.info('Get module FW info: Need to enter password')
-                # Reset password for module using CMIS 4.0
-                self.cdb.module_enter_password(0)
-                fw_info = self._cdb_fw_hdlr.get_firmware_info()
-
-        if fw_info is False or fw_info is None:
-            # Return 0 distinguishes busy/command failure and interface fail from unsupported CDB
-            return {'status': False, 'info': "Failed to get firmware info", 'result': 0}
-
-        fw_status = fw_info.get(cdb_consts.CDB1_FIRMWARE_STATUS, {})
-        ImageARunning = int(fw_status.get(cdb_consts.CDB1_BANKA_OPER_STATUS, False))
-        ImageACommitted = int(fw_status.get(cdb_consts.CDB1_BANKA_ADMIN_STATUS, False))
-        ImageAValid = int(fw_status.get(cdb_consts.CDB1_BANKA_VALID_STATUS, True))
-        ImageBRunning = int(fw_status.get(cdb_consts.CDB1_BANKB_OPER_STATUS, False))
-        ImageBCommitted = int(fw_status.get(cdb_consts.CDB1_BANKB_ADMIN_STATUS, False))
-        ImageBValid = int(fw_status.get(cdb_consts.CDB1_BANKB_VALID_STATUS, True))
-
-        if ImageAValid == 0:
-            ImageA = '{}.{}.{}'.format(
-                fw_info.get(cdb_consts.CDB1_BANKA_MAJOR_VERSION, 0),
-                fw_info.get(cdb_consts.CDB1_BANKA_MINOR_VERSION, 0),
-                fw_info.get(cdb_consts.CDB1_BANKA_BUILD_VERSION, 0)
-            )
-        else:
-            ImageA = "N/A"
-        txt += 'Image A Version: {}\n'.format(ImageA)
-
-        if ImageBValid == 0:
-            ImageB = '{}.{}.{}'.format(
-                fw_info.get(cdb_consts.CDB1_BANKB_MAJOR_VERSION, 0),
-                fw_info.get(cdb_consts.CDB1_BANKB_MINOR_VERSION, 0),
-                fw_info.get(cdb_consts.CDB1_BANKB_BUILD_VERSION, 0)
-            )
-        else:
-            ImageB = "N/A"
-        txt += 'Image B Version: {}\n'.format(ImageB)
-
-        FactoryImage = '{}.{}.{}'.format(
-            fw_info.get(cdb_consts.CDB1_FACTORY_MAJOR_VERSION, 0),
-            fw_info.get(cdb_consts.CDB1_FACTORY_MINOR_VERSION, 0),
-            fw_info.get(cdb_consts.CDB1_FACTORY_BUILD_VERSION, 0)
-        )
-        txt += 'Factory Image Version: {}\n'.format(FactoryImage)
-
-        ActiveFirmware = 'N/A'
-        InactiveFirmware = 'N/A'
-        if ImageARunning == 1:
-            RunningImage = 'A'
-            ActiveFirmware = ImageA
-            InactiveFirmware = ImageB
-        elif ImageBRunning == 1:
-            RunningImage = 'B'
-            ActiveFirmware = ImageB
-            InactiveFirmware = ImageA
-        else:
-            RunningImage = 'N/A'
-        if ImageACommitted == 1:
-            CommittedImage = 'A'
-        elif ImageBCommitted == 1:
-            CommittedImage = 'B'
-        else:
-            CommittedImage = 'N/A'
-
-        txt += 'Running Image: {}\n'.format(RunningImage)
-        txt += 'Committed Image: {}\n'.format(CommittedImage)
-        txt += 'Active Firmware: {}\n'.format(ActiveFirmware)
-        txt += 'Inactive Firmware: {}\n'.format(InactiveFirmware)
-        return {'status': True, 'info': txt, 'result': (ImageA, ImageARunning, ImageACommitted, ImageAValid, ImageB, ImageBRunning, ImageBCommitted, ImageBValid, ActiveFirmware, InactiveFirmware)}
-
-    def cdb_run_firmware(self, mode = 0x01):
-        # run module FW (CMD 0109h)
-        return self.cdb.run_fw_image(mode)
-
-    def cdb_commit_firmware(self):
-        return self.cdb.commit_fw_image()
-
-    def module_fw_run(self, mode = 0x01):
-        """
-        This command is used to start and run a selected image.
-        This command transfers control from the currently
-        running firmware to a selected firmware that is started. It
-        can be used to switch between firmware versions, or to
-        perform a restart of the currently running firmware.
-        mode:
-        00h = Traffic affecting Reset to Inactive Image.
-        01h = Attempt Hitless Reset to Inactive Image
-        02h = Traffic affecting Reset to Running Image.
-        03h = Attempt Hitless Reset to Running Image
-
-        This function returns True if firmware run successfully completes.
-        Otherwise it will return False.
-        """
-        # run module FW (CMD 0109h)
-        txt = ''
-        if self.cdb is None:
-            return False, "CDB NOT supported on this module"
-        starttime = time.time()
-        fw_run_status = self.cdb.run_fw_image(mode)
-        if fw_run_status == 1:
-            txt += 'Module FW run: Success\n'
-        # password issue
-        elif fw_run_status == 70:
-            string = 'Module FW run: Need to enter password\n'
-            logger.info(string)
-            self.cdb.module_enter_password()
-            fw_run_status = self.cdb.run_fw_image(mode)
-            txt += 'FW_run_status %d\n' %fw_run_status
-        else:
-            # self.cdb.abort_fw_download()
-            txt += 'Module FW run: Fail\n'
-            txt += 'FW_run_status %d\n' %fw_run_status
-            return False, txt
-        elapsedtime = time.time()-starttime
-        logger.info('Module FW run time: %.2f s\n' %elapsedtime)
-        logger.info(txt)
-        return True, txt
-
-    def module_fw_commit(self):
-        """
-        The host uses this command to commit the running image
-        so that the module will boot from it on future boots.
-
-        This function returns True if firmware commit successfully completes.
-        Otherwise it will return False.
-        """
-        txt = ''
-        if self.cdb is None:
-            return False, "CDB NOT supported on this module"
-        # commit module FW (CMD 010Ah)
-        starttime = time.time()
-        fw_commit_status= self.cdb.commit_fw_image()
-        if fw_commit_status == 1:
-            txt += 'Module FW commit: Success\n'
-        # password issue
-        elif fw_commit_status == 70:
-            string = 'Module FW commit: Need to enter password\n'
-            logger.info(string)
-            self.cdb.module_enter_password()
-            fw_commit_status = self.cdb.commit_fw_image()
-            txt += 'FW_commit_status %d\n' %fw_commit_status
-        else:
-            # self.cdb.abort_fw_download()
-            txt += 'Module FW commit: Fail\n'
-            txt += 'FW_commit_status %d\n' %fw_commit_status
-            return False, txt
-        elapsedtime = time.time()-starttime
-        logger.info('Module FW commit time: %.2f s\n' %elapsedtime)
-        logger.info(txt)
-        return True, txt
-
-    def cdb_firmware_download_complete(self):
-        # complete FW download (CMD 0107h)
-        return self.cdb.validate_fw_image()
-
-    def cdb_start_firmware_download(self, startLPLsize, startdata, imagesize):
-        return self.cdb.start_fw_download(startLPLsize, bytearray(startdata), imagesize)
-
-    def cdb_lpl_block_write(self, address, data):
-        return self.cdb.block_write_lpl(address, data)
-
-    def cdb_epl_block_write(self, address, data, autopaging_flag, writelength):
-        return self.cdb.block_write_epl(address, data, autopaging_flag, writelength)
-
-    def cdb_enter_host_password(self, password):
-        return self.cdb.module_enter_password(password)
-
-    def module_fw_download(self, startLPLsize, maxblocksize, lplonly_flag, autopaging_flag, writelength, imagepath):
-        """
-        This function performs the download of a firmware image to module eeprom
-        It starts CDB download by writing the header of start header size
-        from the designated firmware file to the local payload page 0x9F, with CDB command 0101h.
-
-        Then it repeatedly reads from the given firmware file and write to the payload
-        space advertised from the first step. We use CDB command 0103h to write to the local payload;
-        we use CDB command 0104h to write to the extended paylaod. This step repeats until it reaches
-        end of the firmware file, or the CDB status failed.
-
-        The last step is to complete the firmware upgrade with CDB command 0107h.
-
-        Note that if the download process fails anywhere in the middle, we need to run CDB command 0102h
-        to abort the upgrade before we restart another upgrade process.
-
-        This function returns True if download successfully completes. Otherwise it will return False where it fails.
-        """
-        txt = ''
-        if self.cdb is None:
-            return False, "CDB NOT supported on this module"
-
-        # start fw download (CMD 0101h)
-        starttime = time.time()
-        try:
-            f = open(imagepath, 'rb')
-        except FileNotFoundError:
-            txt += 'Image path  %s is incorrect.\n' % imagepath
-            logger.info(txt)
-            return False, txt
-
-        f.seek(0, 2)
-        imagesize = f.tell()
-        f.seek(0, 0)
-        startdata = f.read(startLPLsize)
-        logger.info('\nStart FW downloading')
-        logger.info("startLPLsize is %d" %startLPLsize)
-        fw_start_status = self.cdb.start_fw_download(startLPLsize, bytearray(startdata), imagesize)
-        if fw_start_status == 1:
-            string = 'Start module FW download: Success\n'
-            logger.info(string)
-        # password error
-        elif fw_start_status == 70:
-            string = 'Start module FW download: Need to enter password\n'
-            logger.info(string)
-            self.cdb.module_enter_password()
-            self.cdb.start_fw_download(startLPLsize, bytearray(startdata), imagesize)
-        else:
-            string = 'Start module FW download: Fail\n'
-            txt += string
-            self.cdb.abort_fw_download()
-            txt += 'FW_start_status %d\n' %fw_start_status
-            logger.info(txt)
-            return False, txt
-        elapsedtime = time.time()-starttime
-        logger.info('Start module FW download time: %.2f s' %elapsedtime)
-
-        # start periodically writing (CMD 0103h or 0104h)
-        # assert maxblocksize == 2048 or lplonly_flag
-        if lplonly_flag:
-            BLOCK_SIZE = 116
-        else:
-            BLOCK_SIZE = maxblocksize
-        address = 0
-        remaining = imagesize - startLPLsize
-        logger.info("\nTotal size: {} start bytes: {} remaining: {}".format(imagesize, startLPLsize, remaining))
-        while remaining > 0:
-            if remaining < BLOCK_SIZE:
-                count = remaining
-            else:
-                count = BLOCK_SIZE
-            data = f.read(count)
-            if lplonly_flag:
-                fw_download_status = self.cdb.block_write_lpl(address, data)
-            else:
-                fw_download_status = self.cdb.block_write_epl(address, data, autopaging_flag, writelength)
-            if fw_download_status != 1:
-                self.cdb.abort_fw_download()
-                txt += 'CDB download failed. CDB Status: %d\n' %fw_download_status
-                txt += 'FW_download_status %d\n' %fw_download_status
-                logger.info(txt)
-                return False, txt
-            elapsedtime = time.time()-starttime
-            address += count
-            remaining -= count
-            progress = (imagesize - remaining) * 100.0 / imagesize
-            logger.info('Address: {:#08x}; Count: {}; Remain: {:#08x}; Progress: {:.2f}%; Time: {:.2f}s'.format(address, count, remaining, progress, elapsedtime))
-
-        elapsedtime = time.time()-starttime
-        logger.info('Total module FW download time: %.2f s' %elapsedtime)
-
-        time.sleep(2)
-        # complete FW download (CMD 0107h)
-        fw_complete_status = self.cdb.validate_fw_image()
-        if fw_complete_status == 1:
-            string = 'Module FW download complete: Success'
-            logger.info(string)
-            txt += string
-        else:
-            txt += 'Module FW download complete: Fail\n'
-            txt += 'FW_complete_status %d\n' %fw_complete_status
-            logger.info(txt)
-            return False, txt
-        elapsedtime = time.time()-elapsedtime-starttime
-        string = 'Complete module FW download time: %.2f s\n' %elapsedtime
-        logger.info(string)
-        txt += string
-        return True, txt
-
-    def module_fw_upgrade(self, imagepath):
-        """
-        This function performs firmware upgrade.
-        1.  show FW version in the beginning
-        2.  check module advertised FW download capability
-        3.  configure download
-        4.  show download progress
-        5.  configure run downloaded firmware
-        6.  configure commit downloaded firmware
-        7.  show FW version in the end
-
-        imagepath specifies where firmware image file is located.
-        target_firmware is a string that specifies the firmware version to upgrade to
-
-        This function returns True if download successfully completes.
-        Otherwise it will return False.
-        """
-        result = self.get_module_fw_info()
-        try:
-            _, _, _, _, _, _, _, _, _, _ = result['result']
-        except (ValueError, TypeError):
-            return result['status'], result['info']
-        result = self.get_module_fw_mgmt_feature()
-        try:
-            startLPLsize, maxblocksize, lplonly_flag, autopaging_flag, writelength = result['feature']
-        except (ValueError, TypeError):
-            return result['status'], result['info']
-        download_status, txt = self.module_fw_download(startLPLsize, maxblocksize, lplonly_flag, autopaging_flag, writelength, imagepath)
-        if not download_status:
-            return False, txt
-        switch_status, switch_txt = self.module_fw_switch()
-        status = download_status and switch_status
-        txt += switch_txt
-        return status, txt
-
-    def module_fw_switch(self):
-        """
-        This function switch the active/inactive module firmware in the current module memory
-        This function returns True if firmware switch successfully completes.
-        Otherwise it will return False.
-        If not both images are valid, it will stop firmware switch and return False
-        """
-        txt = ''
-        result = self.get_module_fw_info()
-        try:
-            (ImageA_init, ImageARunning_init, ImageACommitted_init, ImageAValid_init,
-             ImageB_init, ImageBRunning_init, ImageBCommitted_init, ImageBValid_init, _, _) = result['result']
-        except (ValueError, TypeError):
-            return result['status'], result['info']
-        if ImageAValid_init == 0 and ImageBValid_init == 0:
-            self.module_fw_run(mode = 0x01)
-            time.sleep(60)
-            self.module_fw_commit()
-            (ImageA, ImageARunning, ImageACommitted, ImageAValid,
-             ImageB, ImageBRunning, ImageBCommitted, ImageBValid, _, _) = self.get_module_fw_info()['result']
-            # detect if image switch happened
-            txt += 'Before switch Image A: %s; Run: %d Commit: %d, Valid: %d\n' %(
-                ImageA_init, ImageARunning_init, ImageACommitted_init, ImageAValid_init
-            )
-            txt += 'Before switch Image B: %s; Run: %d Commit: %d, Valid: %d\n' %(
-                ImageB_init, ImageBRunning_init, ImageBCommitted_init, ImageBValid_init
-            )
-            txt += 'After switch Image A: %s; Run: %d Commit: %d, Valid: %d\n' %(ImageA, ImageARunning, ImageACommitted, ImageAValid)
-            txt += 'After switch Image B: %s; Run: %d Commit: %d, Valid: %d\n' %(ImageB, ImageBRunning, ImageBCommitted, ImageBValid)
-            if (ImageARunning_init == 1 and ImageARunning == 1) or (ImageBRunning_init == 1 and ImageBRunning == 1):
-                txt += 'Switch did not happen.\n'
-                logger.info(txt)
-                return False, txt
-            else:
-                logger.info(txt)
-                return True, txt
-        else:
-            txt += 'Not both images are valid.'
-            logger.info(txt)
-            return False, txt
 
     def get_transceiver_status(self):
         """
