+++ sonic_platform_base/sonic_xcvr/cdb/cdb.py
@@ -11,6 +11,7 @@ from ..xcvr_eeprom import XcvrEeprom
 class CdbCmdHandler(XcvrEeprom):
     def __init__(self, reader, writer, mem_map):
         super(CdbCmdHandler, self).__init__(reader, writer, mem_map)
+        self.last_cmd_status = None
 
     def read_reply(self, cdb_cmd_id):
         """
@@ -63,12 +64,12 @@ class CdbCmdHandler(XcvrEeprom):
         assert timeout > delay, "Timeout must be greater than delay"
 
         while (delay < timeout):
-            time.sleep(cdb_consts.CDB_MAX_CAPTURE_TIME / 1000)
-            delay += cdb_consts.CDB_MAX_CAPTURE_TIME
-
             status = self.read(cdb_consts.CDB1_CMD_STATUS)
+
             if (status is None) or \
                     (True == status[cdb_consts.CDB1_IS_BUSY]):
+                time.sleep(cdb_consts.CDB_MAX_CAPTURE_TIME / 1000)
+                delay += cdb_consts.CDB_MAX_CAPTURE_TIME
                 continue
 
             if (True == status[cdb_consts.CDB1_HAS_FAILED]):
@@ -87,6 +88,7 @@ class CdbCmdHandler(XcvrEeprom):
         """
         Send CDB command, wait for completion and check status
         """
+        self.last_cmd_status = None
         # Write the command to the CDB
         if True != self.write_cmd(cdb_cmd_id, payload):
             print(f"Failed to write CDB command: {cdb_cmd_id}")
@@ -94,6 +96,7 @@ class CdbCmdHandler(XcvrEeprom):
 
         # Wait for the command to complete
         ret, status = self.wait_for_cdb_status(timeout)
+        self.last_cmd_status = status
         if not ret:
             print(f"CDB command: {cdb_cmd_id} failed to complete or read status")
             return None
@@ -117,6 +120,22 @@ class CdbCmdHandler(XcvrEeprom):
         """
         status = self.read(cdb_consts.CDB1_COMMAND_RESULT)
         return status
+
+    def get_cmd_status_code(self):
+        """
+        Get the cached status dict from the last send_cmd call.
+        Returns None if no command was sent or I2C failed.
+        """
+        return self.last_cmd_status
+
+    def enter_password(self, password=cdb_consts.CDB_DEFAULT_PASSWORD):
+        """
+        Enter host password via CDB command 0001h.
+        Default host password is 00001011h.
+        Returns True if password accepted, False/None otherwise.
+        """
+        payload = {"password": password}
+        return self.send_cmd(cdb_consts.CDB_ENTER_PASSWORD_CMD, payload)
     
     def write_lpl_block(self, blkaddr, blkdata):
         """
@@ -127,9 +146,7 @@ class CdbCmdHandler(XcvrEeprom):
             "blkdata" : blkdata
         }
         # Send the CDB write firmware LPL command
-        if True != self.write_cmd(cdb_consts.CDB_WRITE_FIRMWARE_LPL_CMD, payload):
-            status = self.get_last_cmd_status()
-            print(f"Write LPL block status: {status}")
+        return self.send_cmd(cdb_consts.CDB_WRITE_FIRMWARE_LPL_CMD, payload)
 
     def write_epl_pages(self, blkdata):
         """
