+++ sonic_platform_base/sonic_xcvr/fields/cdb_consts.py
@@ -7,6 +7,8 @@ CDB1_IS_BUSY = "Cdb1IsBusy"
 CDB1_HAS_FAILED = "Cdb1HasFailed"
 CDB1_CMD_STATUS_FIELD = "Cdb1CmdStatus"
 CDB1_COMMAND_RESULT ="Cdb1CommandResult"
+CDB_PASSWORD_ERROR_CODE = 0x06
+CDB_DEFAULT_PASSWORD = 0x00001011
 
 
 #Firmware Info
